tiny tiny refactor
