+++ code/internal/handlers/http/products.go
@@ -2,118 +2,17 @@ package http
 
 import (
 	"context"
-	"net/http"
-	"strconv"
-	"time"
-
-	"github.com/SOAT1StackGoLang/tech-challenge/helpers"
-	"github.com/SOAT1StackGoLang/tech-challenge/internal/core/domain"
 	"github.com/SOAT1StackGoLang/tech-challenge/internal/core/ports"
 	restfulspec "github.com/emicklei/go-restful-openapi/v2"
 	"github.com/emicklei/go-restful/v3"
 	"github.com/google/uuid"
+	"net/http"
+	"strconv"
 )
 
-type (
-	ProductsHttpHandler struct {
-		ctx        context.Context
-		productsUC ports.ProductsUseCase
-	}
-
-	Product struct {
-		UpdateProduct
-		InsertionProduct
-		CreatedAt string `json:"created_at,omitempty" readOnly:"true"`
-		UpdatedAt string `json:"updated_at,omitempty" readOnly:"true"`
-		DeletedAt string `json:"deleted_at,omitempty" readOnly:"true"`
-	}
-
-	InsertionProduct struct {
-		UserID      string `json:"user_id,omitempty"`
-		Name        string `json:"name"`
-		Description string `json:"description"`
-		CategoryID  string `json:"category_id"`
-		Price       string `json:"price"`
-	}
-
-	UpdateProduct struct {
-		InsertionProduct
-		ID string `json:"id,omitempty"`
-	}
-
-	ProductList struct {
-		Products []Product `json:"products"`
-		Limit    int       `json:"limit" default:"10"`
-		Offset   int       `json:"offset"`
-		Total    int64     `json:"total"`
-	}
-)
-
-func (p *Product) fromDomain(product *domain.Product) {
-	if p == nil {
-		p = &Product{}
-	}
-
-	price := helpers.ParseDecimalToString(product.Price)
-
-	p.ID = product.ID.String()
-
-	if !product.CreatedAt.IsZero() {
-		p.CreatedAt = product.CreatedAt.Format(time.RFC3339)
-	} else {
-		p.CreatedAt = ""
-	}
-	if !product.UpdatedAt.IsZero() {
-		p.UpdatedAt = product.UpdatedAt.Format(time.RFC3339)
-	} else {
-		p.UpdatedAt = ""
-	}
-	if !product.DeletedAt.IsZero() {
-		p.DeletedAt = product.UpdatedAt.Format(time.RFC3339)
-	} else {
-		p.DeletedAt = ""
-	}
-
-	p.Name = product.Name
-	p.Description = product.Description
-	p.CategoryID = product.CategoryID.String()
-	p.Price = price
-}
-
-func (p *Product) toDomain() *domain.Product {
-	if p == nil {
-		panic("empty product")
-	}
-
-	return domain.ParseProductToDomain(
-		helpers.SafeUUIDFromString(p.ID),
-		helpers.SafeUUIDFromString(p.CategoryID),
-		p.Name,
-		p.Description,
-		p.Price,
-	)
-}
-
-func (iP *InsertionProduct) toDomain() *domain.Product {
-	if iP == nil {
-		panic("empty product")
-	}
-
-	return domain.NewProduct(uuid.New(), helpers.SafeUUIDFromString(iP.CategoryID), iP.Name, iP.Description, iP.Price)
-}
-
-func (uP *UpdateProduct) toDomain() *domain.Product {
-	if uP == nil {
-		panic("empty product")
-	}
-
-	return domain.ParseProductToDomain(
-		helpers.SafeUUIDFromString(uP.ID),
-		helpers.SafeUUIDFromString(uP.CategoryID),
-		uP.Name,
-		uP.Description,
-		uP.Price,
-	)
+type ProductsHttpHandler struct {
+	ctx        context.Context
+	productsUC ports.ProductsUseCase
 }
 
 func NewProductsHttpHandler(ctx context.Context, productsUC ports.ProductsUseCase, ws *restful.WebService) *ProductsHttpHandler {
@@ -124,7 +23,7 @@ func NewProductsHttpHandler(ctx context.Context, productsUC ports.ProductsUseCas
 
 	tags := []string{"products"}
 
-	ws.Route(ws.GET("/products/{id}").To(handler.GetProduct).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
+	ws.Route(ws.GET("/products/{id}").To(handler.handleGetProduct).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
 		Doc("Obtém dados do produto identificado pelo ID fornecido").
 		Param(ws.PathParameter("id", "ID do produto").DataType("string")).
 		Metadata(restfulspec.KeyOpenAPITags, tags).
@@ -132,28 +31,28 @@ func NewProductsHttpHandler(ctx context.Context, productsUC ports.ProductsUseCas
 		Returns(200, "OK", Product{}).
 		Returns(500, "ID de produto não cadastrado ou outro erro", nil))
 
-	ws.Route(ws.POST("/products").To(handler.InsertProduct).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
+	ws.Route(ws.POST("/products").To(handler.handleInsertProduct).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
 		Doc("Cadastra produto").
 		Metadata(restfulspec.KeyOpenAPITags, tags).
 		Reads(InsertionProduct{}). // from the request
 		Returns(200, "Produto cadastrado com sucesso", Product{}).
 		Returns(500, "Erro ao cadastrar produto", nil))
 
-	ws.Route(ws.PUT("/products").To(handler.UpdateProduct).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
+	ws.Route(ws.PUT("/products").To(handler.handleUpdateProduct).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
 		Doc("Atualiza dados do produto").
 		Metadata(restfulspec.KeyOpenAPITags, tags).
 		Reads(UpdateProduct{}). // from the request
 		Returns(200, "Dados do produto atualizados com sucesso", Product{}).
 		Returns(500, "Erro ao atualizar dados do produto", nil))
 
-	ws.Route(ws.DELETE("/products").To(handler.DeleteProduct).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
+	ws.Route(ws.DELETE("/products").To(handler.handleDeleteProduct).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
 		Doc("Remove produto").
 		Metadata(restfulspec.KeyOpenAPITags, tags).
 		Reads(QueryStruct{}). // from the request
 		Returns(200, "Produto removido com sucesso", nil).
 		Returns(500, "Erro ao remover produto", nil))
 
-	ws.Route(ws.GET("/products").To(handler.ListProductsByCategory).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
+	ws.Route(ws.GET("/products").To(handler.handleListProductsByCategory).Consumes(restful.MIME_JSON).Produces(restful.MIME_JSON).
 		Doc("Lista produtos da categoria especificada").
 		Param(ws.QueryParameter("category-id", "ID da categoria").DataType("string")).
 		Param(ws.QueryParameter("limit", "Quantidade máxima de entradas que pode retornar").DataType("string")).
@@ -166,7 +65,7 @@ func NewProductsHttpHandler(ctx context.Context, productsUC ports.ProductsUseCas
 	return handler
 }
 
-func (pH *ProductsHttpHandler) GetProduct(request *restful.Request, response *restful.Response) {
+func (pH *ProductsHttpHandler) handleGetProduct(request *restful.Request, response *restful.Response) {
 	id := request.PathParameter("id")
 
 	uid, err := uuid.Parse(id)
@@ -186,7 +85,7 @@ func (pH *ProductsHttpHandler) GetProduct(request *restful.Request, response *re
 	_ = response.WriteAsJson(prod)
 
 }
-func (pH *ProductsHttpHandler) InsertProduct(request *restful.Request, response *restful.Response) {
+func (pH *ProductsHttpHandler) handleInsertProduct(request *restful.Request, response *restful.Response) {
 	var iProd InsertionProduct
 
 	if err := request.ReadEntity(&iProd); err != nil {
@@ -210,7 +109,7 @@ func (pH *ProductsHttpHandler) InsertProduct(request *restful.Request, response 
 	prod.fromDomain(product)
 	_ = response.WriteAsJson(prod)
 }
-func (pH *ProductsHttpHandler) UpdateProduct(request *restful.Request, response *restful.Response) {
+func (pH *ProductsHttpHandler) handleUpdateProduct(request *restful.Request, response *restful.Response) {
 	var upProd UpdateProduct
 
 	if err := request.ReadEntity(&upProd); err != nil {
@@ -234,7 +133,7 @@ func (pH *ProductsHttpHandler) UpdateProduct(request *restful.Request, response 
 	prod.fromDomain(product)
 	_ = response.WriteAsJson(prod)
 }
-func (pH *ProductsHttpHandler) DeleteProduct(request *restful.Request, response *restful.Response) {
+func (pH *ProductsHttpHandler) handleDeleteProduct(request *restful.Request, response *restful.Response) {
 	var dS QueryStruct
 
 	if err := request.ReadEntity(&dS); err != nil {
@@ -260,7 +159,7 @@ func (pH *ProductsHttpHandler) DeleteProduct(request *restful.Request, response 
 
 	response.WriteHeader(http.StatusOK)
 }
-func (pH *ProductsHttpHandler) ListProductsByCategory(request *restful.Request, response *restful.Response) {
+func (pH *ProductsHttpHandler) handleListProductsByCategory(request *restful.Request, response *restful.Response) {
 	id := request.QueryParameter("category-id")
 	limitS := request.QueryParameter("limit")
 	offsetS := request.QueryParameter("offset")
